Move poseidon lookup to constraint builder
